requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
slowapi==0.1.9
pydantic==2.8.2
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
PyPDF2==3.0.1
requests==2.32.3
streamlit==1.37.1
python-multipart==0.0.9
